+++ src/features/WorkingDirectory/index.ts
@@ -1,0 +1,1 @@
+export { openAddWorkingDirModal } from './AddWorkingDirModal';
